Refactor character routes into modular routers
