+++ server/routes/characters.js
@@ -178,86 +178,6 @@ characterRouter.route('/update-stats/:id').put((req, res, next) => {
   });
 });
 
-// --------------------------------------------------Skills Section------------------------------------------------
-
-// This section will update skills.
-characterRouter.route('/update-skills/:id').put(async  (req, res, next) => {
-  const id = { _id: ObjectId(req.params.id) };
-  const db_connect = req.db;
-  try {
-    const result = await db_connect.collection("Characters").findOneAndUpdate(
-      id,
-      {
-        $set: {
-          "appraise": req.body.appraise,
-          "balance": req.body.balance,
-          "bluff": req.body.bluff,
-          "climb": req.body.climb,
-          "concentration": req.body.concentration,
-          "decipherScript": req.body.decipherScript,
-          "diplomacy": req.body.diplomacy,
-          "disableDevice": req.body.disableDevice,
-          "disguise": req.body.disguise,
-          "escapeArtist": req.body.escapeArtist,
-          "forgery": req.body.forgery,
-          "gatherInfo": req.body.gatherInfo,
-          "handleAnimal": req.body.handleAnimal,
-          "heal": req.body.heal,
-          "hide": req.body.hide,
-          "intimidate": req.body.intimidate,
-          "jump": req.body.jump,
-          "listen": req.body.listen,
-          "moveSilently": req.body.moveSilently,
-          "openLock": req.body.openLock,
-          "ride": req.body.ride,
-          "search": req.body.search,
-          "senseMotive": req.body.senseMotive,
-          "sleightOfHand": req.body.sleightOfHand,
-          "spot": req.body.spot,
-          "survival": req.body.survival,
-          "swim": req.body.swim,
-          "tumble": req.body.tumble,
-          "useTech": req.body.useTech,
-          "useRope": req.body.useRope,
-        },
-      },
-      { returnDocument: 'after' }
-    );
-    res.status(200).json(result.value);
-  } catch (err) {
-    res.status(500).send('Server error');
-  }
-});
-
-// This section will update added skills.
-characterRouter.route('/update-add-skill/:id').put((req, res, next) => {
-  let id = { _id: ObjectId(req.params.id) };
-  let db_connect = req.db;
-  db_connect.collection("Characters").updateOne(id, {$set:{
-  'newSkill': req.body.newSkill
-}}, (err, result) => {
-    if (err) { return next(err); }
-    console.log("character knowledge updated");
-    res.send('user updated sucessfully');
-  });
-});
-
-// This section will update ranks of skills.
-characterRouter.route('/updated-add-skills/:id').put(async  (req, res, next) => {
-  const id = { _id: ObjectId(req.params.id) };
-  const db_connect = req.db;
-  try {
-    const result = await db_connect.collection("Characters").findOneAndUpdate(
-      id,
-      { $set: { 'newSkill': req.body.newSkill } },
-      { returnDocument: 'after' }
-    );
-    res.status(200).json(result.value);
-  } catch (err) {
-    res.status(500).send('Server error');
-  }
-});
-
 // --------------------------------------------------Health Section--------------------------------------------------------
 
 // This section will update tempHealth.
@@ -298,239 +218,6 @@ characterRouter.route('/update-health/:id').put((req, res, next) => {
     console.error(error);
     res.status(500).send('Server error');
   }
-});
-
-// ----------------------------------------------------Weapon Section----------------------------------------------------
-
- // This section will get a list of all the weapons.
- characterRouter.route("/weapons/:campaign").get(function (req, res, next) {
-  let db_connect = req.db;
-  db_connect
-    .collection("Weapons")
-    .find({ campaign: req.params.campaign })
-    .toArray(function (err, result) {
-      if (err) return next(err);
-      res.json(result);
-    });
- });
-
-// This section will update weapons.
-characterRouter.route('/update-weapon/:id').put((req, res, next) => {
-  let id = { _id: ObjectId(req.params.id) };
-  let db_connect = req.db;
-  db_connect.collection("Characters").updateOne(id, {$set:{
-  'weapon': req.body.weapon
-}}, (err, result) => {
-    if (err) { return next(err); }
-    console.log("character weapon updated");
-    res.send('user updated sucessfully');
-  });
-});
-
-// This section will create a new weapon.
-characterRouter.route("/weapon/add").post(function  (req, response, next) {
-  let db_connect = req.db;
-  let myobj = {
-  campaign: req.body.campaign,
-  weaponName: req.body.weaponName,
-  enhancement: req.body.enhancement,
-  damage: req.body.damage,
-  critical: req.body.critical,
-  weaponStyle: req.body.weaponStyle,
-  range: req.body.range
-  };
-  db_connect.collection("Weapons").insertOne(myobj, function (err, res) {
-    if (err) return next(err);
-    response.json(res);
-  });
- });
-// -----------------------------------------------------Armor Section--------------------------------------------------------
-
-// This section will get a list of all the armor.
-characterRouter.route("/armor/:campaign").get(function (req, res, next) {
-  let db_connect = req.db;
-  db_connect
-    .collection("Armor")
-    .find({ campaign: req.params.campaign })
-    .toArray(function (err, result) {
-      if (err) return next(err);
-      res.json(result);
-    });
- });
-
-// This section will create a new armor.
-characterRouter.route("/armor/add").post(function  (req, response, next) {
-  let db_connect = req.db;
-  let myobj = {
-  campaign: req.body.campaign,
-  armorName: req.body.armorName,
-  armorBonus: req.body.armorBonus,
-  maxDex: req.body.maxDex,
-  armorCheckPenalty: req.body.armorCheckPenalty
-  };
-  db_connect.collection("Armor").insertOne(myobj, function (err, res) {
-    if (err) return next(err);
-    response.json(res);
-  });
- });
-
-// This section will update armors.
-characterRouter.route('/update-armor/:id').put((req, res, next) => {
-  let id = { _id: ObjectId(req.params.id) };
-  let db_connect = req.db;
-  db_connect.collection("Characters").updateOne(id, {$set:{
-  'armor': req.body.armor
-}}, (err, result) => {
-    if (err) { return next(err); }
-    console.log("character armor updated");
-    res.send('user updated sucessfully');
-  });
-});
-// ------------------------------------------------------Item Section-----------------------------------------------------------
-
-// This section will get a list of all the items.
-characterRouter.route("/items/:campaign").get(function (req, res, next) {
-  let db_connect = req.db;
-  db_connect
-    .collection("Items")
-    .find({ campaign: req.params.campaign })
-    .toArray(function (err, result) {
-      if (err) return next(err);
-      res.json(result);
-    });
- });
-
-// This section will create a new item.
-characterRouter.route("/item/add").post(function  (req, response, next) {
-  let db_connect = req.db;
-  let myobj = {
-    campaign: req.body.campaign,
-    itemName: req.body.itemName, 
-    notes: req.body.notes,
-    str: req.body.str,
-    dex: req.body.dex,
-    con: req.body.con,
-    int: req.body.int,
-    wis: req.body.wis,
-    cha: req.body.cha,
-    appraise: req.body.appraise,
-    balance: req.body.balance,
-    bluff: req.body.bluff,
-    climb: req.body.climb,
-    concentration: req.body.concentration,
-    decipherScript: req.body.decipherScript,
-    diplomacy: req.body.diplomacy,
-    disableDevice: req.body.disableDevice,
-    disguise: req.body.disguise,
-    escapeArtist: req.body.escapeArtist,
-    forgery: req.body.forgery,
-    gatherInfo: req.body.gatherInfo,
-    handleAnimal: req.body.handleAnimal,
-    heal: req.body.heal,
-    hide: req.body.hide,
-    intimidate: req.body.intimidate,
-    jump: req.body.jump,
-    listen: req.body.listen,
-    moveSilently: req.body.moveSilently,
-    openLock: req.body.openLock,
-    ride: req.body.ride,
-    search: req.body.search,
-    senseMotive: req.body.senseMotive,
-    sleightOfHand: req.body.sleightOfHand,
-    spot: req.body.spot,
-    survival: req.body.survival,
-    swim: req.body.swim,
-    tumble: req.body.tumble,
-    useTech: req.body.useTech,
-    useRope: req.body.useRope
-  };
-  db_connect.collection("Items").insertOne(myobj, function (err, res) {
-    if (err) return next(err);
-    response.json(res);
-  });
- });
-
- // This section will update items.
-characterRouter.route('/update-item/:id').put((req, res, next) => {
-  let id = { _id: ObjectId(req.params.id) };
-  let db_connect = req.db;
-  db_connect.collection("Characters").updateOne(id, {$set:{
-  'item': req.body.item
-}}, (err, result) => {
-    if (err) { return next(err); }
-    console.log("character item updated");
-    res.send('user updated sucessfully');
-  });
-});
-
-// ------------------------------------------------------Feat Section-----------------------------------------------------------
-
-// This section will get a list of all the feats.
-characterRouter.route("/feats").get(function (req, res, next) {
-  let db_connect = req.db;
-  db_connect
-    .collection("Feats")
-    .find({})
-    .toArray(function (err, result) {
-      if (err) return next(err);
-      res.json(result);
-    });
- });
-
-// This section will create a new feat.
-characterRouter.route("/feat/add").post(function  (req, response, next) {
-  let db_connect = req.db;
-  let myobj = {
-    featName: req.body.featName, 
-    notes: req.body.notes,
-    appraise: req.body.appraise,
-    balance: req.body.balance,
-    bluff: req.body.bluff,
-    climb: req.body.climb,
-    concentration: req.body.concentration,
-    decipherScript: req.body.decipherScript,
-    diplomacy: req.body.diplomacy,
-    disableDevice: req.body.disableDevice,
-    disguise: req.body.disguise,
-    escapeArtist: req.body.escapeArtist,
-    forgery: req.body.forgery,
-    gatherInfo: req.body.gatherInfo,
-    handleAnimal: req.body.handleAnimal,
-    heal: req.body.heal,
-    hide: req.body.hide,
-    intimidate: req.body.intimidate,
-    jump: req.body.jump,
-    listen: req.body.listen,
-    moveSilently: req.body.moveSilently,
-    openLock: req.body.openLock,
-    ride: req.body.ride,
-    search: req.body.search,
-    senseMotive: req.body.senseMotive,
-    sleightOfHand: req.body.sleightOfHand,
-    spot: req.body.spot,
-    survival: req.body.survival,
-    swim: req.body.swim,
-    tumble: req.body.tumble,
-    useTech: req.body.useTech,
-    useRope: req.body.useRope
-  };
-  db_connect.collection("Feats").insertOne(myobj, function (err, res) {
-    if (err) return next(err);
-    response.json(res);
-  });
- });
-
- // This section will update feats.
-characterRouter.route('/update-feat/:id').put((req, res, next) => {
-  let id = { _id: ObjectId(req.params.id) };
-  let db_connect = req.db;
-  db_connect.collection("Characters").updateOne(id, {$set:{
-  'feat': req.body.feat
-}}, (err, result) => {
-    if (err) { return next(err); }
-    console.log("character feat updated");
-    res.send('user updated sucessfully');
-  });
 });
 
 // --------------------------------------------------------Level Up Section--------------------------------------------------------------------
