+++ server/routes/index.js
@@ -7,6 +7,9 @@ const auth = require('./auth');
 const users = require('./users');
 const campaigns = require('./campaigns');
 const characters = require('./characters');
+const skills = require('./skills');
+const feats = require('./feats');
+const equipment = require('./equipment');
 
 routes.use(async (req, res, next) => {
   try {
@@ -21,5 +24,8 @@ auth(routes);
 users(routes);
 campaigns(routes);
 characters(routes);
+skills(routes);
+feats(routes);
+equipment(routes);
 
 module.exports = routes;
